#28: Extract theme definitions into separate file
