+++ src/theme/emotion.d.ts
@@ -1,7 +1,7 @@
 import '@emotion/react'
-
-import { CustomTheme } from './selectors'
 import { Theme as EmotionTheme } from '@emotion/react'
+
+import { CustomTheme } from './theme'
 
 declare module '@emotion/react' {
   export interface Theme extends EmotionTheme, CustomTheme {}
