+++ src/theme/selectors.ts
@@ -2,168 +2,7 @@ import { createSelector } from '@reduxjs/toolkit'
 
 import { RootState } from '../store'
 
-const SPACING_UNIT = 8
-const ANIMATION_DURATION_SHORT = 0.1
-const ANIMATION_DURATION_LONG = 0.4
-
-const TOOLBAR_WIDTH = 305
-const PAGE_WIDTH = 794 // A4 width = 8.27 × 96 (DPI) ≈ 794px
-const PAGE_HEIGHT = 1123 // A4 height = 11.69 × 96 (DPI) ≈ 1123px
-const PAGE_PADDING = 75 // 20mm at 96dpi = 0.787402 * 96 = 75.590592px
-const PAGE_PADDING_PRINT = 20 // 20mm
-
-export const TOOLBAR_BREAKPOINT = PAGE_WIDTH + 2 * TOOLBAR_WIDTH
-
-const breakpoints = {
-  sm: '@media screen and (min-width: 0px)',
-  md: '@media screen and (min-width: 768px)',
-  page: `@media screen and (min-width: ${PAGE_WIDTH}px)`,
-  toolbar: `@media screen and (min-width: ${TOOLBAR_BREAKPOINT}px)`,
-  lg: '@media screen and (min-width: 1024px)',
-  xl: '@media screen and (min-width: 1440px)',
-}
-
-interface BaseTheme {
-  breakpoints: {
-    sm: string
-    md: string
-    page: string
-    toolbar: string
-    lg: string
-    xl: string
-  }
-  animations: {
-    interaction: string
-    transition: string
-    stepsDuration: string
-    tickStepsDuration: string
-  }
-  interactions: {
-    hoverScale: number
-    activeScale: number
-    hoverOpacity: number
-    activeOpacity: number
-  }
-  spacing: (multiplier: number) => string
-  layout: {
-    toolbarWidth: string
-    pageWidth: string
-    pageHeight: string
-    pagePadding: string
-    pagePaddingPrint: string
-  }
-  fontSize: {
-    large: string
-    normal: string
-    small: string
-    tiny: string
-    editor: string
-    kbd: string
-  }
-}
-
-export type CustomTheme = BaseTheme & {
-  mode: 'light' | 'dark'
-  colors: {
-    primary: string
-    secondary: string
-    title: string
-    text: string
-    link: string
-    backdrop: string
-    paper: string
-    background: string
-    page: string
-    shadow: string
-    modal: string
-    modalBackdrop: string
-    punctuation: string
-    editorFocus: string
-  }
-  opacity: {
-    disabled: number
-  }
-}
-
-const BASE_THEME: BaseTheme = {
-  breakpoints,
-  animations: {
-    interaction: `${ANIMATION_DURATION_SHORT}s ease-in-out`,
-    transition: `${ANIMATION_DURATION_LONG}s ease-in-out`,
-    stepsDuration: `${ANIMATION_DURATION_LONG}s`,
-    tickStepsDuration: `${ANIMATION_DURATION_SHORT * 2}s`,
-  },
-  interactions: {
-    hoverScale: 1.05,
-    activeScale: 0.97,
-    hoverOpacity: 1,
-    activeOpacity: 0.8,
-  },
-  spacing: (multiplier: number) => `${multiplier * SPACING_UNIT}px`,
-  layout: {
-    toolbarWidth: `${TOOLBAR_WIDTH}px`,
-    pageWidth: `${PAGE_WIDTH}px`,
-    pageHeight: `${PAGE_HEIGHT}px`,
-    pagePadding: `${PAGE_PADDING}px`,
-    pagePaddingPrint: `${PAGE_PADDING_PRINT}mm !important`,
-  },
-  fontSize: {
-    large: '1.5rem',
-    normal: '1rem',
-    small: '0.875rem',
-    tiny: '0.75rem',
-    editor: '1rem',
-    kbd: '0.6rem',
-  },
-}
-
-const LIGHT_THEME: CustomTheme = {
-  ...BASE_THEME,
-  mode: 'light',
-  colors: {
-    primary: 'rgba(156, 109, 255, 1)',
-    secondary: 'rgba(55, 55, 55, 0.5)',
-    title: 'rgba(55, 55, 55, 1)',
-    text: 'rgba(55, 55, 55, 1)',
-    link: 'rgba(55, 55, 55, 1)',
-    backdrop: 'rgba(0, 0, 0, 0.075)',
-    paper: 'rgba(156, 109, 255, 0.1)',
-    background: 'rgb(248, 248, 248)',
-    page: 'rgba(255, 255, 255, 1)',
-    shadow: 'rgba(0, 0, 0, 0.1)',
-    modal: 'rgba(255, 255, 255, 1)',
-    modalBackdrop: 'rgba(255, 255, 255, 0.8)',
-    punctuation: 'rgba(156, 109, 255, 0.5)',
-    editorFocus: 'rgba(156, 109, 255, 0.3)',
-  },
-  opacity: {
-    disabled: 0.5,
-  },
-}
-
-const DARK_THEME: CustomTheme = {
-  ...BASE_THEME,
-  mode: 'dark',
-  colors: {
-    primary: 'rgba(156, 109, 255, 1)', // Keeping primary color unchanged
-    secondary: 'rgba(200, 200, 200, 0.5)',
-    title: 'rgba(255, 255, 255, 1)',
-    text: 'rgba(255, 255, 255, 0.87)', // Text is generally lighter in dark themes
-    link: 'rgba(255, 255, 255, 0.87)', // Same as text
-    backdrop: 'rgba(255, 255, 255, 0.1)', // A lighter backdrop for dark theme
-    paper: 'rgba(156, 109, 255, 0.2)', // A slightly more prominent shade for paper in dark theme
-    background: 'rgba(34, 34, 34, 1)', // Dark grey background color
-    page: 'rgba(44, 44, 44, 1)',
-    shadow: 'rgba(255, 255, 255, 0.05)',
-    modal: 'rgba(44, 44, 44, 1)',
-    modalBackdrop: 'rgba(34, 34, 34, 0.9)',
-    punctuation: 'rgba(156, 109, 255, 0.5)',
-    editorFocus: 'rgba(156, 109, 255, 0.3)',
-  },
-  opacity: {
-    disabled: 0.5,
-  },
-}
+import { DARK_THEME, LIGHT_THEME } from './theme'
 
 export const selectTheme = (s: RootState) => s.theme
